app/requirements.txt
flask
flask-login
flask-wtf
flask-injector
flask-jwt
flask_mysqldb
email_validator
alembic
gunicorn
Faker
tarantool
redis
python-redis-cache
celery
blinker
flask-socketio
pika
